+++ src/auth_handler.rs
@@ -1,15 +1,20 @@
 use actix::{Handler, Message};
 use actix_identity::Identity;
 use actix_web::FromRequest;
-use actix_web::{dev::Payload, Error, HttpRequest};
+use actix_web::{dev::Payload, web, Error, HttpRequest, HttpResponse, Responder, ResponseError};
 use bcrypt::verify;
 use diesel::prelude::*;
 
 use crate::errors::ServiceError;
 use crate::models::{DbExecutor, SlimUser, User};
 use crate::utils::decode_token;
+use actix::Addr;
+use futures::Future;
+
+use crate::utils::create_token;
 
 #[derive(Debug, Deserialize)]
+
 pub struct AuthData {
     pub email: String,
     pub password: String,
@@ -57,4 +62,30 @@ impl FromRequest for LoggedUser {
         }
         Err(ServiceError::Unauthorized.into())
     }
+}
+
+pub fn login(
+    auth_data: web::Json<AuthData>,
+    id: Identity,
+    db: web::Data<Addr<DbExecutor>>,
+) -> impl Future<Item = HttpResponse, Error = Error> {
+    db.send(auth_data.into_inner())
+        .from_err()
+        .and_then(move |res| match res {
+            Ok(user) => {
+                let token = create_token(&user)?;
+                id.remember(token);
+                Ok(HttpResponse::Ok().into())
+            }
+            Err(err) => Ok(err.error_response()),
+        })
+}
+
+pub fn logout(id: Identity) -> impl Responder {
+    id.forget();
+    HttpResponse::Ok()
+}
+
+pub fn get_me(logged_user: LoggedUser) -> HttpResponse {
+    HttpResponse::Ok().json(logged_user)
 }
